// packages/templates/builder/nodes.d.ts
export interface BuilderNodeInterface {
    render(): string;
}
export declare class BuilderNodesContainer<T extends BuilderNodeInterface> implements BuilderNodeInterface {
    delimiter: string;
    private nodes;
    constructor(nodes?: Array<T>, delimiter?: string);
    add(node: T | string): void;
    addList(nodes: Array<T>): void;
    replace(container: BuilderNodesContainer<T>): void;
    isEmpty(): boolean;
    render(): string;
}
// ...
export declare class BuilderClass implements BuilderNodeInterface {
    name: string;
    args: Array<string>;
    beforeClass: BuilderNodesContainer<BuilderNodeInterface>;
    afterClass: BuilderNodesContainer<BuilderNodeInterface>;
    body: BuilderNodesContainer<BuilderNodeInterface>;
    methods: BuilderNodesContainer<BuilderMethod>;
    constructor(name: string, args?: Array<string>);
    render(): string;
}
export declare function createMethod(parent: BuilderClass, name: string, args?: Array<string>, setup?: (method: BuilderMethod) => void): BuilderMethod;
export declare class BuilderMethod implements BuilderNodeInterface {
    parent: BuilderClass;
    name: string;
    args: Array<string>;
    body: BuilderNodesContainer<BuilderNodeInterface>;
    end: BuilderNodesContainer<BuilderNodeInterface>;
    constructor(parent: BuilderClass, name: string, args?: Array<string>);
    render(): string;
}
export declare function createTemplateMethod(parent: BuilderClass, id: number, setup?: (method: BuilderTemplateMethod) => void): BuilderTemplateMethod;
export declare class BuilderTemplateMethod extends BuilderMethod {
    id: number;
    constructor(parent: BuilderClass, id: number);
}